Update to Pulls and Scan tool

+++ pages/views/scan.py
@@ -16,11 +16,20 @@ from django.http import JsonResponse
 # pdf parsing
 from io import StringIO
 from pdfminer.high_level import *
+
+# cfts settings
+from cfts import settings as cftsSettings
+
+# db models
+from pages.models import *
+
+# regular expressions
+import re
 # ====================================================================
 
 
 @login_required
-def scan(request):
+def scan(request,pullZip):
     # request context
     rc = {
         'bodyText': 'Scan Tool'
@@ -28,20 +37,28 @@ def scan(request):
 
     # POST
     if request.method == 'POST':
-        form_files = request.FILES
-##    print( form_files )
-
-        for i, f in enumerate(form_files.getlist("toScan")):
-            zf = ZipFile(f)
+        if pullZip !="none":
+            pullZip = os.path.join(cftsSettings.BASE_DIR,"pulls",pullZip)
+            zf = ZipFile(pullZip)
             zf.extractall(settings.SCANTOOL_DIR)
             results = runScan()
+
+        else:
+            form_files = request.FILES
+            for i, f in enumerate(form_files.getlist("toScan")):
+                zf = ZipFile(f)
+                zf.extractall(settings.SCANTOOL_DIR)
+                results = runScan()
 
         # clean up after yourself
         cleanup(settings.SCANTOOL_DIR)
         return JsonResponse(results, safe=False)
 
     # GET
-    return render(request, 'pages/scan.html', {'rc': rc})
+    if pullZip !="none":
+        return render(request, 'pages/scan.html', {'rc': rc, 'pullZip': pullZip })
+    else:
+        return render(request, 'pages/scan.html', {'rc': rc})
 
 
 def runScan():
@@ -52,37 +69,43 @@ def runScan():
 
     # \cfts\scan should contain all the user folders from the zip file
     for root, subdirs, files in os.walk(scan_dir):
+        txt = re.compile('_email(\d+)?.txt')
+        eml = re.compile('_email(\d+)?.eml')
+
         for filename in files:
-            file_results = None
-            file_path = os.path.join(root, filename)
-##      print( '\t- file %s (full path: %s)' % ( filename, file_path ) )
-            temp, ext = os.path.splitext(filename)
+            if txt.match(filename) == None and eml.match(filename) == None:
+                file_results = None
+                file_path = os.path.join(root, filename)
+    ##      print( '\t- file %s (full path: %s)' % ( filename, file_path ) )
+                temp, ext = os.path.splitext(filename)
 
-            if(ext in office_filetype_list):
-                file_results = scanOfficeFile(file_path)
-            elif(ext == '.pdf'):
-                textFile = open(root+"\\"+temp+".txt", "w")
-                pdf2Text = StringIO()
-                with open(file_path, 'rb') as pdf:
-                    extract_text_to_fp(
-                        pdf, pdf2Text, output_type='text', codec='utf-8')
-                    textFile.write(pdf2Text.getvalue().strip())
-                    textFile.close()
+                if(ext in office_filetype_list):
+                    file_results = scanOfficeFile(file_path)
+                elif(ext == '.pdf'):
+                    textFile = open(root+"\\"+temp+".txt", "w",encoding='utf-8')
+                    pdf2Text = StringIO()
+                    with open(file_path, 'rb',) as pdf:
+                        extract_text_to_fp(
+                            pdf, pdf2Text, output_type='text', codec='utf-8')
+                        textFile.write(pdf2Text.getvalue().strip())
+                        textFile.close()
 
-                text_path = os.path.join(root+"\\"+temp+".txt")
-                file_results = scanFile(text_path)
-                file_results['file'] = file_path
-                os.remove(text_path)
+                    text_path = os.path.join(root+"\\"+temp+".txt")
+                    file_results = scanFile(text_path)
+                    if file_results is not None:
+                        file_results['file'] = file_path
+                    os.remove(text_path)
 
-            else:
-                file_results = scanFile(file_path)
+                else:
+                    file_results = scanFile(file_path)
 
-            if(file_results is not None):
-                result = {}
-                result['file'] = file_path
-                result['found'] = file_results
-                scan_results.append(result)
-
+                if(file_results is not None):
+                    result = {}
+                    result['file'] = file_path
+                    result['found'] = file_results
+                    scan_results.append(result)
+                
+            
     return scan_results
 
 
@@ -109,24 +132,27 @@ def scanOfficeFile(office_file):
 
 
 def scanFile(text_file):
-    result = {
-        'file': text_file,
-        'findings': []
-    }
+    # result = {
+    #     'file': text_file,
+    #     'findings': []
+    # }
+    result = None
 
-    dirty_word_list = ["SECRET", "S//", "NOFORN",
-                       "C//", "CONFIDENTIAL", "PRV", "LVY"]
     reg_lst = []
-    for raw_regex in dirty_word_list:
-        reg_lst.append(re.compile(raw_regex))
+
+    for raw_regex in DirtyWord.objects.filter(case_sensitive = False):
+        reg_lst.append(re.compile(raw_regex.word, re.IGNORECASE))
+    
+    for raw_regex in DirtyWord.objects.filter(case_sensitive = True):
+        reg_lst.append(re.compile(raw_regex.word))
 
     try:
         with open(text_file, "r", encoding="utf-8") as f:
             f_content = f.read()
-            # result = {
-            #    'file': text_file,
-            #    'findings': []
-            # }
+            result = {
+               'file': text_file,
+               'findings': []
+            }
             for compiled_reg in reg_lst:
                 found = re.finditer(compiled_reg, f_content)
                 for match in found:
@@ -134,7 +160,7 @@ def scanFile(text_file):
                     result['findings'].append(
                         'This file contains the term: %s' % (match.group()))
             if not len(result['findings']):
-                result['findings'].append("Scan returned no dirty words.")
+                result = None
 
     except UnicodeDecodeError:
         # Found non-text data
